Add data from ME 452 final project
Add swept sine data and one step response test in the directory
July_07_2009.  This data is being used for my ME 452 summer class, but
it is also good data that excites the second mode of the beam.  There
is a problem with coherence at the peak of the first mode that I am
not sure I understand right now.  It seems to get worse with smaller
amplitudes, so it may be friction related.